requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
python-dotenv
